Restore typed event decoding
