+++ app/ts/simulation/logHandlers.ts
@@ -7,7 +7,6 @@ import type { ParsedEvent, TokenVisualizerResult } from '../types/EnrichedEthere
 import { decodeEventStrict } from '../utils/abiRuntime.js'
 
 const isBigintArray = (value: unknown): value is readonly bigint[] => Array.isArray(value) && value.every((entry) => typeof entry === 'bigint')
-const isRecord = (value: unknown): value is Record<string, unknown> => typeof value === 'object' && value !== null && !Array.isArray(value)
 
 export function handleERC20TransferLog(eventLog: EthereumEvent): TokenVisualizerResult[] {
 	if (eventLog.topics[1] === undefined || eventLog.topics[2] === undefined) throw new Error('unknown log')
@@ -87,7 +86,6 @@ export function handleERC1155TransferBatch(eventLog: EthereumEvent): TokenVisual
 		topics: eventLog.topics.map((topic) => EthereumBytes32.serialize(topic)).filter((topic): topic is `0x${ string }` => topic !== undefined),
 	})
 	if (parsed.eventName !== 'TransferBatch') throw new Error('Malformed ERC1155 TransferBatch Event')
-	if (!isRecord(parsed.args)) throw new Error('Malformed ERC1155 TransferBatch Event')
 	const { _ids, _values } = parsed.args
 	if (!isBigintArray(_ids) || !isBigintArray(_values)) throw new Error('Malformed ERC1155 TransferBatch Event')
 	return [...Array(_ids.length)].map((_, index) => {
